Add basic GCP support for network egress verification
